+++ Order.hpp
@@ -24,7 +24,7 @@ private:
 	Side side;
 	int quantity;
 	int limit_price;
-	int reserved_cash;
+	long long reserved_cash;
 	Status status;
 
 public:
@@ -35,11 +35,11 @@ public:
 	Side get_side() const;
 	int get_quantity() const;
 	int get_limit_price() const;
-	int get_reserved_cash() const;
+	long long get_reserved_cash() const;
 	Status get_status() const;
 	// const std::unordered_map<std::string, int>& get_reserved_stocks() const;
 	void set_quantity(int new_quantity);
-	void set_reserved_cash(int new_reserved_cash);
+	void set_reserved_cash(long long new_reserved_cash);
 	// void add_reserved_stocks(const std::unordered_map<std::string, int>& new_reserved_stocks);
 	// void remove_reserved_stocks(const std::unordered_map<std::string, int>& stocks_to_remove);
 	void set_status(Status new_status);
